+++ src/main/java/com/x1/frans/supplier/query/service/SupplierQueryService.java
@@ -11,5 +11,5 @@ public interface SupplierQueryService {
 
     List<SupplierListDTO> getAllSuppliers();
 
-    SupplierDetailDTO getSupplierDetail(int supplierId);
+    SupplierDetailDTO getSupplierDetail(long supplierId);
 }
