+++ src/main/java/com/x1/frans/supplier/query/service/SupplierQueryServiceImpl.java
@@ -29,7 +29,7 @@ public class SupplierQueryServiceImpl implements SupplierQueryService {
     }
 
     @Override
-    public SupplierDetailDTO getSupplierDetail(int supplierId) {
+    public SupplierDetailDTO getSupplierDetail(long supplierId) {
         return supplierQueryMapper.getSupplierDetail(supplierId);
     }
 }
